Reuploaded the code with lot's of new features and a brand new architecture
New Features:
1- The program now is dynamic and can extract the a, b, and c coefficients from any equation given to him
Ex: 2X + 5x - 3 => a = 2, b = 5, c = -3. X - 2x + X + 5 - 3x - 2 => a = 2, b = -5, c = 3
2- The program also writes the solutions of the equation as they are written in your notebook. and not as a fraction anymore 
For example the solutions of the equation X + 4 - 6 = 0. will be written as -4 + 2sqrt(10) / 2 , -4 - 2sqrt(10) / 2, and not 1.1622, -5.1622
3- The program will simplify the roots, for example, if sqrt(18) does exist in the answer, it will write it as 3sqrt(2)

Also, the architecture of the code has been changed completely, there is no code in the controls any more, instead, it's put into different classes according to what is the job of the code. And in the controls, it's just a matter of calling these classes.

Finally, the way that the code works changed from expecting the equation to be in a certain form in order to extract a, b, and c. To expecting an equation in any form and the data extracting algorithm will work on them all